Implementing the empty-typeahead directive.
This directive has the following requirements:

1) The controller of the HTML page that contains typeahead elements,
must import EmptyTypeaheadService, and expose the following methods
to the scope:

$scope.fieldComparator = EmptyTypeaheadService.fieldComparator;
$scope.onFocus = EmptyTypeaheadService.onFocus;

2) The input field of the HTML template with the typeahead directive
must contain the following two directives:

ng-focus="onFocus($event)" empty-typeahead

3) The typeahead directive of the input element must be modified by
adding the fieldComparator filter, for example:

typeahead="f for f in fields | filter:$viewValue:fieldComparator"

As a proof of concept, the asset name element is currently using this
feature.

The idea is apply this feature over all the typeahead fields, and
replace the select fields with typeahead fields (because they look
nicer).